+++ examples/basic-nested-store/src/pages/About/store.ts
@@ -1,0 +1,8 @@
+import { createStore } from 'ice';
+import modelAbout from './model';
+
+const store = createStore({
+  about: modelAbout
+});
+
+export default store;