+++ examples/with-rax/src/pages/Home/index.tsx
@@ -1,5 +1,5 @@
 import { createElement } from 'rax';
-import { usePageShow, usePageHide } from 'rax-app';
+import { usePageShow, usePageHide, getSearchParams } from 'rax-app';
 import View from 'rax-view';
 import Text from 'rax-text';
 import Logo from '@/components/Logo';
@@ -8,6 +8,11 @@ import './index.css';
 
 export default function Home(props) {
   const { history } = props;
+
+  const searchParams = getSearchParams();
+
+  console.log('home search params =>', searchParams);
+  console.log('home page props =>', props);
 
   usePageShow(() => {
     console.log('home show...');
@@ -20,9 +25,18 @@ export default function Home(props) {
   return (
     <View className="home">
       <Logo />
-      <Text className="title">Welcome to Your Rax App!!!</Text>
-      <Text className="info">More information about Rax</Text>
-      <Text className="info" onClick={() => history.push('/about')}>Go About</Text>
+      <Text className="title">{props?.data?.title || 'Welcome to Your Rax App'}</Text>
+      <Text className="info">{props?.data?.info || 'More information about Rax'}</Text>
+      <Text className="info" onClick={() => history.push('/about?id=1')}>Go About</Text>
     </View>
   );
 }
+
+Home.getInitialProps = async () => {
+  return {
+    data: {
+      title: 'Welcome to Your Rax App with SSR',
+      info: 'More information about Rax'
+    }
+  };
+};