+++ packages/create-app-container/src/createNavigation.ts
@@ -45,6 +45,7 @@ const activatePageComponent = (route, { createElement }) => {
     });
 };
 
+// eslint-disable-next-line
 const renderAlivePages = (routes, { alivePages, currentPage, isAlivePage, createElement }) => {
   if (alivePages.length > 0) {
     return createElement(
@@ -82,61 +83,63 @@ const renderAlivePages = (routes, { alivePages, currentPage, isAlivePage, create
   return null;
 };
 
-const createNavigation = (api) => (props) => {
-  const { createElement, useEffect, useState, Fragment } = api;
-  const { staticConfig, component, history, routes } = props;
-  const { maxAlivePageNum, tabBar } = staticConfig;
-
-  const [, setUpdateTemp] = useState(null);
-
-  const Component = component;
-  const currentPathname = history.location.pathname;
-  const currentPage = routes.find(route => route.path === currentPathname) || {};
-
-  const isAlivePage = currentPage.keepAlive;
-  useEffect(() => {
-    // Use display control alive page, need get alive page list.
-    routes.forEach((route) => {
-      if (route.keepAlive) {
-        alivePages.push(route);
-      }
-    });
-    // If current page is alive page, need update routes.
-    if (isAlivePage) {
-      // @ts-ignore
-      _updatePageTrigger(Date.now());
-    }
-  }, []);
-
-  // Props to page component
-  const pageProps = {};
-  Object.keys(props).forEach((key) => {
-    if (key !== 'staticConfig' && key !== 'component') {
-      pageProps[key] = props[key];
-    }
-  });
-
-  config.pageProps = pageProps;
-  config.routes = routes;
-  _updatePageTrigger = setUpdateTemp;
-  // eslint-disable-next-line
-  maxAlivePageNum && (config.maxAlivePageNum = maxAlivePageNum);
-
+const createNavigation = (api) => {
   const TabBar = createTabBar(api);
 
-  return createElement(
-    Fragment,
-    null,
-    isAlivePage ? null : createElement(Component, pageProps),
-    renderAlivePages(routes, { alivePages, currentPage, isAlivePage, createElement }),
-    createElement(
-      TabBar,
-      {
-        config: tabBar,
-        history
+  return (props) => {
+    const { createElement, useEffect, useState, Fragment } = api;
+    const { staticConfig, component, history, routes } = props;
+    const { maxAlivePageNum, tabBar } = staticConfig;
+
+    const [, setUpdateTemp] = useState(null);
+
+    const Component = component;
+    const currentPathname = history.location.pathname;
+    const currentPage = routes.find(route => route.path === currentPathname) || {};
+
+    const isAlivePage = currentPage.keepAlive;
+    useEffect(() => {
+      // Use display control alive page, need get alive page list.
+      routes.forEach((route) => {
+        if (route.keepAlive) {
+          alivePages.push(route);
+        }
+      });
+      // If current page is alive page, need update routes.
+      if (isAlivePage) {
+        // @ts-ignore
+        _updatePageTrigger(Date.now());
       }
-    )
-  );
+    // eslint-disable-next-line
+    }, []);
+
+    // Props to page component
+    const pageProps = {};
+    Object.keys(props).forEach((key) => {
+      if (key !== 'staticConfig' && key !== 'component') {
+        pageProps[key] = props[key];
+      }
+    });
+
+    config.pageProps = pageProps;
+    config.routes = routes;
+    _updatePageTrigger = setUpdateTemp;
+    // eslint-disable-next-line
+    maxAlivePageNum && (config.maxAlivePageNum = maxAlivePageNum);
+    return createElement(
+      Fragment,
+      null,
+      isAlivePage ? null : createElement(Component, pageProps),
+      renderAlivePages(routes, { alivePages, currentPage, isAlivePage, createElement }),
+      createElement(
+        TabBar,
+        {
+          config: tabBar,
+          history
+        }
+      )
+    );
+  };
 };
 
 export default createNavigation;