+++ packages/create-app-shared/src/getSearchParams.ts
@@ -1,0 +1,10 @@
+import * as queryString from 'query-string';
+import { getHistory } from './history';
+
+export default function() {
+  const history = getHistory();
+  if (history && history.location && history.location.search) {
+    return queryString.parse(history.location.search);
+  }
+  return {};
+}